refactor: repositoryのtypeを再エクスポート
